frontend: update kubernetes proxy client
